[11.25/Redapanda] 그룹 지원 및 그룹 가입 API, 스케줄 추가 API 등 생성

+++ BE/WhenWhere/src/main/java/com/example/whenwhere/Dto/ScheduleDto.java
@@ -9,7 +9,8 @@ import java.time.LocalDateTime;
 @ToString
 @Builder
 public class ScheduleDto {
-    private Long id;
+
+    private int id;
 
     private String title;
