+++ BE/WhenWhere/src/main/java/com/example/whenwhere/Entity/Authority.java
@@ -14,5 +14,5 @@ public class Authority {
 
     @Id
     @Column(name="authority_name")
-    private Long authorityName;
+    private Integer authorityName;
 }
